\name{hg18_len}
\alias{hg18_len}
\docType{data}
\title{
A table with information on chromosome lengths for the hg18 human genome build.
}
\description{
The table is used to create genomic windows for the whole chromosome lengths.
}
\usage{data(hg18_len)}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{chr}}{A factor with levels \code{1} \code{10} \code{11} \code{12} \code{13} \code{14} \code{15} \code{16} \code{17} \code{18} \code{19} \code{2} \code{20} \code{21} \code{22} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{X} \code{Y}.}
    \item{\code{length}}{A numeric vector.}
  }
}
\examples{
data(hg18_len)
}
\keyword{datasets}
