\name{seqCNA.annot-package}
\alias{seqCNA.annot-package}
\alias{seqCNA.annot}
\docType{package}
\title{
Annotation for the copy number analysis of deep sequencing cancer data with seqCNA
}
\description{
Provides annotation on GC content, mappability and genomic features for various genomes
}
\details{
\tabular{ll}{
Package: \tab seqCNA.annot\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2013-03-27\cr
License: \tab GPL-3\cr
}
}
\author{
David Mosen-Ansorena

Maintainer: David Mosen-Ansorena <dmosen.gn@cicbiogune.es>
}
