\name{supported.builds}
\alias{supported.builds}
\title{
Names of the genome builds for which the package contains annotation.
}
\description{
A vector with the names of the genome builds with annotation in the package.
}
\usage{
supported.builds()
}
\value{
A vector with the names of the genome builds with annotation in the package.
}
\author{
David Mosen-Ansorena
}
\examples{
supported.builds()
}
\keyword{ Information }
